services/etl-service/requirements.txt
pymongo>=4.0.0
psycopg2-binary>=2.9.0
